clean up routing and make scaffolding for aboutMe Cooldown and Guide components
