+++ src/components/AboutMe.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const AboutMe = () => {
+    return (
+        <div>
+            
+        </div>
+    )
+}
+
+export default AboutMe
