+++ src/components/Cooldown.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Cooldown = () => {
+    return (
+        <div>
+            
+        </div>
+    )
+}
+
+export default Cooldown
