+++ src/components/Guide.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Guide = () => {
+    return (
+        <div>
+            
+        </div>
+    )
+}
+
+export default Guide
