+++ src/styles/footer.css
@@ -1,5 +1,5 @@
 .footerContainer {
-  margin: 0;
-  background-color: var(--pale-blue);
-  height: 5vh;
+    margin: 0;
+    background-color: var(--pale-blue);
+    min-height: 5vh;
 }
