+++ src/styles/header.css
@@ -1,24 +1,24 @@
 .headerContainer {
-  /*this is my reminder for a horizontally divided div
+    /*this is my reminder for a horizontally divided div
   background-image: linear-gradient(var(--pale-blue) 65%, White 35%);*/
-  background-color: var(--pale-blue);
-  height: 20vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
+    background-color: var(--pale-blue);
+    min-height: 20vh;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
 }
 
 .claim {
-  /*border: 1px solid red;*/
-  padding: 0;
-  margin: 0;
-  font-size: 5em;
-  color: var(--deep-purple);
-  font-family: "Codystar", cursive;
+    /*border: 1px solid red;*/
+    padding: 0;
+    margin: 0;
+    font-size: 5em;
+    color: var(--deep-purple);
+    font-family: "Codystar", cursive;
 }
 
 h3 {
-  font-family: "Montserrat", sans-serif;
-  color: var(--deep-purple);
-  padding-bottom: 1rem;
+    font-family: "Montserrat", sans-serif;
+    color: var(--deep-purple);
+    padding-bottom: 1rem;
 }
